Upgrade counters for sites and volume from int to size_t

+++ generic/io_lat_utils.c
@@ -973,7 +973,7 @@ int read_v3_gauge_hdr(gauge_file *gf, int parallel, int *byterevflag)
 	  ny = gh->dims[1];
 	  nz = gh->dims[2];
 	  nt = gh->dims[3];
-	  volume = nx*ny*nz*nt;
+	  volume = (size_t)nx*ny*nz*nt;
 	}
     }
   /* Header byte length for this file */
@@ -1088,7 +1088,7 @@ int read_1996_gauge_hdr(gauge_file *gf, int parallel, int *byterevflag)
 	  ny = gh->dims[1];
 	  nz = gh->dims[2];
 	  nt = gh->dims[3];
-	  volume = nx*ny*nz*nt;
+	  volume = (size_t)nx*ny*nz*nt;
 	}
     }
   
@@ -1258,7 +1258,7 @@ int read_fnal_gauge_hdr(gauge_file *gf, int parallel, int *byterevflag)
 	  ny = gh->dims[1];
 	  nz = gh->dims[2];
 	  nt = gh->dims[3];
-	  volume = nx*ny*nz*nt;
+	  volume = (size_t)nx*ny*nz*nt;
 	}
     }
   
@@ -1456,7 +1456,7 @@ int read_gauge_hdr(gauge_file *gf, int parallel)
 	  ny = gh->dims[1];
 	  nz = gh->dims[2];
 	  nt = gh->dims[3];
-	  volume = nx*ny*nz*nt;
+	  volume = (size_t)nx*ny*nz*nt;
 	}
     }
 
@@ -1498,7 +1498,7 @@ void write_site_list(FILE *fp, gauge_header *gh)
 {
   off_t offset;
   int i;
-  int buf_length;
+  size_t buf_length;
   register site *s;
   int32type coords, *cbuf;
 
@@ -1538,7 +1538,7 @@ void write_site_list(FILE *fp, gauge_header *gh)
       buf_length++;
     }
 
-    if( (int)g_write(cbuf,sizeof(int32type),sites_on_node,fp) != sites_on_node)
+    if( (size_t)g_write(cbuf,sizeof(int32type),sites_on_node,fp) != sites_on_node)
       {
 	printf("write_site_list: Node %d coords write error %d\n",
 	       this_node,errno);fflush(stdout);terminate(1);
